# perf: Optimize readBlock function for large inputs