@klauspost 2024-10-28
@cosnicolaou Could you check out the linter. It seems to be spitting out errors for fairly common truncation:

```
Running [/home/runner/golangci-lint-1.61.0-linux-amd64/golangci-lint run --out-format=github-actions]  ....
::high file=internal/bitstream/bits.go,line=72,col=21::G115: integer overflow conversion int -> uint8 (gosec)
  ::high file=internal/bitstream/bits.go,line=73,col=21::G115: integer overflow conversion int -> uint8 (gosec)
  ::high file=internal/bitstream/bits.go,line=78,col=66::G115: integer overflow conversion int -> uint8 (gosec)
  ::high file=internal/bzip2/bit_reader.go,line=73,col=14::G115: integer overflow conversion int -> uint (gosec)
  ::high file=internal/bzip2/bit_reader.go,line=78,col=12::G115: integer overflow conversion uint64 -> int (gosec)
  ::high file=internal/bzip2/block.go,line=44,col=63::G115: integer overflow conversion int -> uint (gosec)
  ::high file=internal/bzip2/bzip2.go,line=234,col=25::G115: integer overflow conversion uint64 -> uint32 (gosec)
  ::high file=internal/bzip2/bzip2.go,line=287,col=27::G115: integer overflow conversion uint64 -> uint32 (gosec)
  ::high file=internal/bzip2/bzip2.go,line=294,col=17::G115: integer overflow conversion int -> uint (gosec)
  ::high file=internal/bzip2/bzip2.go,line=518,col=21::G115: integer overflow conversion int -> uint32 (gosec)
  ::high file=internal/bzip2/bzip2.go,line=5[33](https://github.com/cosnicolaou/pbzip2/actions/runs/11551720023/job/32149422526?pr=41#step:4:35),col=16::G115: integer overflow conversion int -> uint32 (gosec)
  ::high file=internal/bzip2/huffman.go,line=46,col=16::G115: integer overflow conversion uint64 -> uint16 (gosec)
  ::high file=internal/bzip2/huffman.go,line=51,col=16::G115: integer overflow conversion int -> uint16 (gosec)
  ::high file=internal/bzip2/huffman.go,line=99,col=26::G115: integer overflow conversion int -> uint16 (gosec)
  ::high file=internal/bzip2/huffman.go,line=212,col=20::G115: integer overflow conversion int -> uint16 (gosec)
  ```