@cosnicolaou 2024-10-28
These are caused by the G115 check added to gosec, which still (and likely always will) report some false positives. I'll merge this change and fix up the linting separately.